# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Stratis Marketing Engine","description":"Marketing campaign management with AI capabilities","version":"2.0.0"},"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user.","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login and get access token.","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User","description":"Get current user info.","operationId":"get_current_user_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Logout (client-side token removal).","operationId":"logout_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/campaigns/export/csv":{"get":{"tags":["Export"],"summary":"Export Campaigns Csv","description":"Export all campaigns for current user as CSV.","operationId":"export_campaigns_csv_campaigns_export_csv_get","responses":{"200":{"description":"Successful Response"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/campaigns":{"get":{"tags":["Campaigns"],"summary":"Get Campaigns","description":"Get all campaigns for current user with optional filtering.","operationId":"get_campaigns_campaigns_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","default":"start_date","title":"Sort By"}},{"name":"sort_order","in":"query","required":false,"schema":{"type":"string","default":"ASC","title":"Sort Order"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CampaignResponse"},"title":"Response Get Campaigns Campaigns Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Campaigns"],"summary":"Create Campaign","description":"Create a new campaign.","operationId":"create_campaign_campaigns_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/campaigns/{campaign_id}":{"get":{"tags":["Campaigns"],"summary":"Get Campaign","description":"Get a specific campaign by ID.","operationId":"get_campaign_campaigns__campaign_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"campaign_id","in":"path","required":true,"schema":{"type":"integer","title":"Campaign Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Campaigns"],"summary":"Update Campaign","description":"Update an existing campaign.","operationId":"update_campaign_campaigns__campaign_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"campaign_id","in":"path","required":true,"schema":{"type":"integer","title":"Campaign Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Campaigns"],"summary":"Delete Campaign","description":"Delete a campaign.","operationId":"delete_campaign_campaigns__campaign_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"campaign_id","in":"path","required":true,"schema":{"type":"integer","title":"Campaign Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/campaigns/{campaign_id}/duplicate":{"post":{"tags":["Campaigns"],"summary":"Duplicate Campaign","description":"Duplicate an existing campaign (creates a draft copy).","operationId":"duplicate_campaign_campaigns__campaign_id__duplicate_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"campaign_id","in":"path","required":true,"schema":{"type":"integer","title":"Campaign Id"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/brief":{"post":{"tags":["AI"],"summary":"Generate Brief","description":"Generate a campaign brief using AI.","operationId":"generate_brief_ai_brief_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BriefRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BriefResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/ai/insights":{"post":{"tags":["AI"],"summary":"Generate Insights","description":"Generate portfolio insights using AI.","operationId":"generate_insights_ai_insights_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/dashboard/stats":{"get":{"tags":["Dashboard"],"summary":"Get Dashboard Stats","description":"Get dashboard statistics for current user.","operationId":"get_dashboard_stats_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardStats"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/analytics/stats":{"get":{"tags":["Analytics"],"summary":"Get Analytics","description":"Get comprehensive analytics for current user.","operationId":"get_analytics_analytics_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"tags":["Health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"BriefRequest":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"target_audience":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Target Audience"},"budget":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Budget"},"currency":{"anyOf":[{"type":"string","pattern":"^(USD|EUR|GBP)$"},{"type":"null"}],"title":"Currency","default":"USD"}},"type":"object","required":["name"],"title":"BriefRequest","description":"Request schema for AI brief generation.","example":{"budget":50000,"currency":"USD","name":"Q4 Product Launch","target_audience":"Tech-savvy professionals aged 25-40"}},"BriefResponse":{"properties":{"description":{"type":"string","title":"Description"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"notes":{"type":"string","title":"Notes"}},"type":"object","required":["description","tags","notes"],"title":"BriefResponse","description":"Response schema for AI brief generation.","example":{"description":"Q4 Product Launch campaign targeting tech-savvy professionals.","notes":"Key metrics: 25% increase in brand awareness, 15% conversion rate.","tags":["product-launch","q4-2024","professional"]}},"CampaignCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"budget":{"type":"number","exclusiveMinimum":0.0,"title":"Budget"},"spent":{"type":"number","minimum":0.0,"title":"Spent","default":0.0},"currency":{"allOf":[{"$ref":"#/components/schemas/Currency"}],"default":"USD"},"start_date":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$","title":"Start Date"},"end_date":{"anyOf":[{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"type":"null"}],"title":"End Date"},"target_audience":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Target Audience"},"status":{"allOf":[{"$ref":"#/components/schemas/CampaignStatus"}],"default":"Draft"},"owner":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Owner"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":20},{"type":"null"}],"title":"Tags"},"assets":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Assets"},"notes":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Notes"}},"type":"object","required":["name","budget","start_date"],"title":"CampaignCreate"},"CampaignResponse":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"budget":{"type":"number","exclusiveMinimum":0.0,"title":"Budget"},"spent":{"type":"number","minimum":0.0,"title":"Spent","default":0.0},"currency":{"allOf":[{"$ref":"#/components/schemas/Currency"}],"default":"USD"},"start_date":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$","title":"Start Date"},"end_date":{"anyOf":[{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"type":"null"}],"title":"End Date"},"target_audience":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Target Audience"},"status":{"allOf":[{"$ref":"#/components/schemas/CampaignStatus"}],"default":"Draft"},"owner":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Owner"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":20},{"type":"null"}],"title":"Tags"},"assets":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Assets"},"notes":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Notes"},"id":{"type":"integer","title":"Id"},"budget_usd":{"type":"number","title":"Budget Usd"},"is_expired":{"type":"boolean","title":"Is Expired"},"remaining":{"type":"number","title":"Remaining"},"progress_pct":{"type":"number","title":"Progress Pct"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["name","budget","start_date","id","budget_usd","is_expired","remaining","progress_pct"],"title":"CampaignResponse"},"CampaignStatus":{"type":"string","enum":["Draft","Active","Paused","Completed"],"title":"CampaignStatus","description":"Campaign status enumeration."},"CampaignUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"budget":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Budget"},"spent":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Spent"},"currency":{"anyOf":[{"$ref":"#/components/schemas/Currency"},{"type":"null"}]},"start_date":{"anyOf":[{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"type":"null"}],"title":"End Date"},"target_audience":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Target Audience"},"status":{"anyOf":[{"$ref":"#/components/schemas/CampaignStatus"},{"type":"null"}]},"owner":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Owner"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array","maxItems":20},{"type":"null"}],"title":"Tags"},"assets":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Assets"},"notes":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Notes"}},"type":"object","title":"CampaignUpdate"},"Currency":{"type":"string","enum":["USD","EUR","GBP"],"title":"Currency","description":"Currency enumeration."},"DashboardStats":{"properties":{"active_budget_usd":{"type":"number","title":"Active Budget Usd","description":"Total active budget in USD"},"total_budget_usd":{"type":"number","title":"Total Budget Usd","description":"Total budget across all campaigns in USD"},"total_spent_usd":{"type":"number","title":"Total Spent Usd","description":"Total spent across all campaigns in USD"},"counts_by_status":{"additionalProperties":{"type":"integer"},"type":"object","title":"Counts By Status","description":"Campaign counts by status"},"expired_count":{"type":"integer","title":"Expired Count","description":"Number of expired campaigns"},"total_campaigns":{"type":"integer","title":"Total Campaigns","description":"Total number of campaigns"}},"type":"object","required":["active_budget_usd","total_budget_usd","total_spent_usd","counts_by_status","expired_count","total_campaigns"],"title":"DashboardStats","description":"Dashboard statistics schema."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InsightsRequest":{"properties":{"campaigns":{"items":{"type":"object"},"type":"array","minItems":1,"title":"Campaigns"}},"type":"object","required":["campaigns"],"title":"InsightsRequest","description":"Request schema for AI portfolio insights.","example":{"campaigns":[{"budget":50000,"is_expired":false,"name":"Q4 Product Launch","progress_pct":50.0,"spent":25000,"status":"Active"}]}},"InsightsResponse":{"properties":{"insights":{"type":"string","title":"Insights"}},"type":"object","required":["insights"],"title":"InsightsResponse","description":"Response schema for AI portfolio insights.","example":{"insights":"Portfolio is healthy with 3 active campaigns. One campaign is approaching 90% spend. Recommend reallocating budget from underperforming channels."}},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserRegister":{"properties":{"email":{"type":"string","maxLength":100,"title":"Email"},"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"password":{"type":"string","minLength":8,"title":"Password"},"full_name":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Full Name"}},"type":"object","required":["email","username","password"],"title":"UserRegister"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"username":{"type":"string","title":"Username"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","email","username","is_active"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}